Sanitize inputs, cast IDs, and escape SQL
Harden input handling across multiple modules by casting request IDs and numeric hidden fields to integers, trimming inputs, and applying mysqli_real_escape_string before building queries. Normalize file/attachment handling and replace raw $_POST checks with safer post()/trim() usage. Add identifier/column validation and select-list normalization in searchData.php to prevent SQL injection in dynamic queries. Changes span customer, finance, shopee, stock, users, login, and search scripts to improve security and data consistency.

+++ finance/cred_inv_create.php
@@ -8,7 +8,7 @@ $pageTitle = getPinGroupNameById($connect, $currentPagePin);
 
 $tblName = CRED_NOTES_INV;
 //Current Page Action And Data ID
-$dataId = !empty(input('id')) ? input('id') : post('id');
+$dataId = !empty(input('id')) ? (int) input('id') : (int) post('id');
 
 //Page Redirect Link , Clean LocalStorage , Error Alert Msg 
 $redirectPage = $SITEURL . '/finance/cred_notes_inv_table.php';
